% Generated by roxygen2 (4.0.1): do not edit by hand
\name{clusterSE}
\alias{clusterSE}
\title{Compute clustered standard errors.}
\usage{
clusterSE(fit, cluster.var, data)
}
\arguments{
\item{fit}{A model fit with \code{plm}.}

\item{cluster.var}{A character string naming the grouping variable.}

\item{data}{A data frame containing \code{cluster.var}.}
}
\value{
Output from \code{coeftest} (\pkg{lmtest}) but with clustered standard errors.
}
\description{
Compute clustered standard errors.
}
\examples{
clusterSE(fit, cluster.var, data)
}
\author{
Jonah Gabry <jsg2201@columbia.edu>
}

