% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggMyTS.R
\name{ggMyTS}
\alias{ggMyTS}
\title{Easily plot time series trends with ggplot}
\usage{
ggMyTS(df, varlist, line = TRUE, point = TRUE, pointsize = 3,
  linewidth = 1.25, ...)
}
\arguments{
\item{df}{The data frame in which to look for variables to be plotted. 
Typically created with \code{\link{meltMyTS}}.}

\item{varlist}{A character string or vector naming the variable(s) in
\code{df} to plot. If \code{varlist} is not specified, then all variables
in \code{df} will be used.}

\item{line}{Should lines be plotted? Defaults to \code{TRUE}.}

\item{point}{Should points be plotted? Defaults to \code{TRUE}.}

\item{pointsize}{Size of the points, if \code{point} is \code{TRUE}.}

\item{linewidth}{Width of the line(s), if \code{line} is \code{TRUE}.}

\item{...}{Arguments passed to \code{\link[ggplot2]{geom_line}} and
\code{\link[ggplot2]{geom_point}}.}
}
\value{
A ggplot object that can be further customized using functions
  in the \pkg{ggplot2} package.
}
\description{
\code{ggMyTS} is primarily designed to be used with a data frame created with
\code{\link{meltMyTS}}.
}
\examples{
\dontrun{
ts.data <- ts(cbind(year = 1950:1999, x1 = rnorm(50, sd = 1), x2 = rnorm(50, sd = 1)))
plot.dat <- meltMyTS(ts.data, time.var = "year")
ggMyTS(plot.dat, varlist = c("x1", "x2")) + 
   ggplot2::scale_x_continuous(breaks = seq(1950,1999, by = 5))
}

}
\seealso{
\code{\link{meltMyTS}}
}

