% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Librarian.R
\name{Librarian}
\alias{Librarian}
\title{Load multiple packages with one command}
\usage{
Librarian(pkgs)
}
\arguments{
\item{pkgs}{A character vector of package names.}
}
\description{
Load multiple packages with one command
}
\examples{
\dontrun{
my.packages <- c("ggplot2", "MASS", "plm")
Librarian(my.packages)
}

}
\seealso{
\code{\link[base]{library}}
}

