% Generated by roxygen2 (4.0.1): do not edit by hand
\name{meltMyTS}
\alias{meltMyTS}
\title{Use \pkg{reshape2}'s \code{melt} on multivariate time series object}
\usage{
meltMyTS(mv.ts.object, time.var, keep.vars)
}
\arguments{
\item{mv.ts.object}{a multivariate time series object created with \code{\link[stats]{ts}}.}

\item{time.var}{A character string naming the time variable.
to either side of the cutpoint for which the estimate is calculated.}

\item{keep.vars}{An optional character vector with names of variables to keep.
Variables not named in \code{keep.vars} will be dropped. If \code{keep.vars} is
not specified then all variables in \code{mv.ts.object} will kept.}
}
\value{
A molten data frame.
}
\description{
\code{meltMyTS} is primarily designed to prepare data to be used with
\code{\link[QMSS]{ggMyTS}} (\pkg{QMSS}) to make plots of time series trends
with \code{\link[ggplot2]{ggplot}}.
}
\examples{
See examples in documentation for \\code{\\link[QMSS]{ggMyTS}}.
}
\author{
Jonah Gabry <jsg2201@columbia.edu>. See \code{\link[reshape2]{melt}}
in (\pkg{reshape2}) for the author of the original \code{melt} function.
}
\seealso{
\code{\link[QMSS]{ggMyTS}}, \code{\link[reshape2]{melt.data.frame}}
}

