% Generated by roxygen2 (4.0.1): do not edit by hand
\name{propOddsTest}
\alias{propOddsTest}
\title{Test proportional odds assumption for ordinal logit models}
\usage{
propOddsTest(fit, relaxed.fit)
}
\arguments{
\item{fit}{An ordinal logit model fit with \code{\link[VGAM]{vglm}} (\pkg{VGAM})
with \code{family = propodds}.}

\item{relaxed.fit}{An ordinal logit model fit with \code{\link[VGAM]{vglm}} (\pkg{VGAM})
with \code{family = cumulative(reverse = TRUE)}}
}
\value{
An object of class "htest".
}
\description{
Test proportional odds assumption for ordinal logit models
}
\examples{
propOddsTest(fit, relaxed.fit)
}
\author{
Jonah Gabry <jsg2201@columbia.edu>
}
\seealso{
\code{\link[VGAM]{vglm}}
}

