% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QMSS-package.r
\docType{package}
\name{QMSS-package}
\alias{QMSS}
\alias{QMSS-package}
\title{Companion package for Greg Eirich's QMSS courses at Columbia University.}
\description{
Simple functions, data sets, example code, etc, for QMSS G4015 Quantitative
Analytic Techniques graduate course at Columbia University taught 
by Greg Eirich.
}
\author{
Jonah Gabry <jsg2201@columbia.edu>
}

