% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDplot.R
\name{RDplot}
\alias{RDplot}
\title{Plot regression discontinuity}
\usage{
RDplot(x, gran = 400, bins = 100, which = 1, range, col = c("black",
  "black"), pts = TRUE, xlab = "", ylab = "",
  main = "Plot of Regression Discontinuity", ...)
}
\arguments{
\item{x}{A model fit with \code{\link[rdd]{RDestimate}} (\pkg{rdd})}

\item{gran}{The granularity of the plot. This specifies the number of points 
to either side of the cutpoint for which the estimate is calculated.}

\item{bins}{If the dependent variable is binary, include the number of bins 
within which to average}

\item{which}{Identifies which of the available plots to display. For a sharp
design, the only possibility is 1, the plot of the running variable against
the outcome variable. For a fuzzy design, an additional plot, 2, may also
be displayed, showing the relationship between the running variable and the
treatment variable. Both plots may be displayed with which=c(1,2)}

\item{range}{The range of values of the running variable for which to plot.
This should be a vector of length two of the format c(min,max). To plot
from the minimum to the maximum value, simply enter c("min","max"). The
default is a window 20 times wider than the first listed bandwidth from the
rd object, truncated by the min/max values of the running variable from the
data.}

\item{col}{Colors to distinguish before and after the discontinuity.}

\item{pts}{Should points be plotted. Defaults to TRUE.}

\item{xlab}{Label for the horizontal axis}

\item{ylab}{Label for the vertical axis}

\item{main}{Title for the plot}
}
\description{
\code{RDplot} adds additional features to \code{\link[rdd]{plot.RD}} from the
\pkg{rdd} package.
}
\seealso{
\code{\link[rdd]{plot.RD}}
}

