% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverseCode.R
\name{reverseCode}
\alias{reverseCode}
\title{Reverse-code a variable}
\usage{
reverseCode(var)
}
\arguments{
\item{var}{A numeric or factor variable.}
}
\value{
A reverse-coded version of \code{var}.
}
\description{
Reverse-code a variable
}
\examples{
x <- c(1,1,1,2,3,3)
reverseCode(x)

}

