% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverseCode.R
\name{ReverseThis}
\alias{ReverseThis}
\title{ReverseThis (Deprecated)}
\usage{
ReverseThis(var)
}
\arguments{
\item{var}{A numeric or factor variable.}
}
\description{
Use \code{\link{reverseCode}} instead.
}

