% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sigmaRho}
\alias{sigmaRho}
\title{Compute sigma_u, sigma_e, and rho after fixed or random effects model}
\usage{
sigmaRho(fit)
}
\arguments{
\item{fit}{A fixed effects model fit with \code{\link{plm}} (\pkg{plm}) and
\code{model = "within"} or \code{model = "random"}}
}
\value{
A numeric vector containing the named components \itemize{
\item{\code{sigma_u}}{ Error due to differences between units.}
\item{\code{sigma_e}}{ Error due to differences within units.}
\item{\code{rho}}{ Proportion of variance due to unit effects.}
}. Results also printed to the console.
}
\description{
Compute sigma_u, sigma_e, and rho after fixed or random effects model
}
\examples{
\dontrun{
# Model from plm help page:
data("Produc", package = "plm")
fit <- plm(log(gsp) ~ log(pcap) + log(pc) + log(emp) + unemp,
           data = Produc, index = c("state","year"), model = "random")
sigmaRho(fit)
}
}
\author{
Jonah Gabry <jsg2201@columbia.edu>
}
\seealso{
\code{\link{plm}}
}

