% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stdCoef}
\alias{stdCoef}
\title{Compute standardized regression coefficients for a linear model}
\usage{
stdCoef(fit)
}
\arguments{
\item{fit}{A model fit with \code{\link[stats]{lm}}.}
}
\value{
A named vector containing the standardized (beta) coefficients for \code{fit}.
}
\description{
Compute standardized regression coefficients for a linear model
}
\examples{
lm.realinc <- lm(realinc ~ age + educ, data = GSS_2010)
stdCoef(lm.realinc)
}
\author{
Jonah Gabry <jsg2201@columbia.edu>
}

