% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tab.R
\name{Tab}
\alias{Tab}
\title{Simple table of counts and percentages}
\usage{
Tab(var, digits = 2, useNA = c("no", "ifany", "always"))
}
\arguments{
\item{var}{A variable.}

\item{digits}{An integer indicating the number of decimals places to be used
for rounding.}

\item{useNA}{Should \code{NA} values be tabulated? Defaults to \code{"no"}.}
}
\value{
A table of counts, percentages, and cumulative percentages for levels
  of \code{var}.
}
\description{
Simple table of counts and percentages
}
\details{
If \code{useNA = "ifany"} then \code{NA} values will be included if
  any exist. If \code{useNA = "always"} then if \code{var} does not contain
  any \code{NA} values a count, percentage, and cumulative percentage of 0
  will be reported.
}
\examples{
x <- c(1,1,1,2,3,3,NA,NA)
y <- c(1,1,1,2,3,3)
Tab(x)
Tab(x, useNA = "ifany")
Tab(y, useNA = "always")
with(GSS_2010, Tab(partyid, digits = 3, useNA = "ifany"))

}
\seealso{
\code{\link[stats]{xtabs}}, \code{\link[base]{table}}, 
  \code{\link[base]{prop.table}}
}

