% Generated by roxygen2 (4.0.1): do not edit by hand
\name{TabX}
\alias{TabX}
\title{Cross Tabulation}
\usage{
TabX(x, y, digits = 3, ...)
}
\arguments{
\item{x,y}{The variables for the cross tabulation.}

\item{digits}{Number of digits for rounding proportions.}
}
\description{
\code{XTab} is a modified version of \code{\link{CrossTable}} (\pkg{gmodels}).
}
\examples{
data("GSS_2010", package = "QMSS")
with(GSS_2010, TabX(sex, race))
}
\author{
Jonah Gabry <jsg2201@columbia.edu> See \code{\link{CrossTable}} for the authors
of the function on which \code{XTab} is based.
}
\seealso{
\code{\link{plot.RD}}
}

