% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterSE.R
\name{clusterSE}
\alias{clusterSE}
\title{Clustered standard errors}
\usage{
clusterSE(fit, cluster.var, data)
}
\arguments{
\item{fit}{A model fit with \code{\link[plm]{plm}}.}

\item{cluster.var}{A character string naming the grouping/cluster variable.}

\item{data}{A data frame containing \code{cluster.var} Only needed if 
\code{cluster.var} is not included in \code{index}. See 'Examples' below.}
}
\value{
Output from \code{\link[lmtest]{coeftest}} but with clustered
  standard errors.
}
\description{
Clustered standard errors
}
\note{
\code{clusterSE} does not work with models fit with \code{lm}, however
  a similar model can be fit with \code{\link[plm]{plm}} using the option
  \code{model = "pooling"}. You can then use \code{clusterSE} to compute
  clustered standard errors and retest the coefficients.
}
\examples{
# Model from plm help page:
data("Produc", package = "plm")
fit <- plm(log(gsp) ~ log(pcap) + log(pc) + log(emp) + unemp,
           data = Produc, index = c("state","year"), model = "random")

clusterSE(fit, cluster.var = "state") # don't need data argument since "state" is included in index 

}
\seealso{
\code{\link[lmtest]{coeftest}}, \code{\link[plm]{plm}}
}

