% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ggMyTS}
\alias{ggMyTS}
\title{Easily plot time series trends with \code{\link[ggplot2]{ggplot}}}
\usage{
ggMyTS(df, varlist, ...)
}
\arguments{
\item{df}{The data frame in which to look for variables to be plotted. Typically
created with \code{\link[QMSS]{meltMyTS}} (\pkg{QMSS})}

\item{varlist}{A character string or vector naming the variable(s) in \code{df} to plot.}

\item{...}{Other options that will be passed to both \code{\link[ggplot2]{geom_line}}
and \code{\link[ggplot2]{geom_point}}. See examples.}
}
\value{
A \code{ggplot} object.
}
\description{
\code{ggMyTS} is primarily designed to be used with a data frame created with
\code{\link[QMSS]{meltMyTS}} (\pkg{QMSS}).
}
\examples{
\dontrun{
keep.vars <- c("year", "n.confinan", "fulltime")
plot.dat <- meltMyTS(mv.ts.object = by.year.ts,
                     time.var = "year", keep.vars = keep.vars)
ggMyTS(plot.dat, varlist = c("n.confinan", "fulltime"))
ggMyTS(plot.dat, "n.confinan", color = "forestgreen")
}
}
\author{
Jonah Gabry <jsg2201@columbia.edu>. See \code{\link[ggplot2]{ggplot}}
for the author of the \code{ggplot} function.
}
\seealso{
\code{\link[QMSS]{meltMyTS}}
}

