% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Librarian}
\alias{Librarian}
\title{Load multiple packages with one command}
\usage{
Librarian(pkgs)
}
\arguments{
\item{pkgs}{A character vector of package names.}
}
\description{
Load multiple packages with one command
}
\details{
The \code{Librarian} function loads the packages named in \code{pkgs}.
}
\examples{
my.packages <- c("car", "foreign")
Librarian(my.packages)
}
\author{
Jonah Gabry <jsg2201@columbia.edu>
}
\seealso{
\code{\link[base]{library}}
}

