% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meltMyTS.R
\name{meltMyTS}
\alias{meltMyTS}
\title{Use \pkg{reshape2}'s \code{melt} on multivariate time series object}
\usage{
meltMyTS(mv.ts.object, time.var, keep.vars)
}
\arguments{
\item{mv.ts.object}{A multivariate time series object created with
\code{\link[stats]{ts}}.}

\item{time.var}{A character string naming the time variable.}

\item{keep.vars}{An optional character vector with names of variables to
keep. If \code{keep.vars} is not specified then all variables in
\code{mv.ts.object} will be kept. However, if any variables are named then
all other variables will be dropped, except \code{time.var}, which is
always kept.}
}
\value{
A molten data frame.
}
\description{
\code{meltMyTS} is primarily designed to prepare data to be used with 
\code{\link{ggMyTS}} to make plots of time series trends with
\code{\link[ggplot2]{ggplot}}.
}
\examples{
# See examples in documentation for ggMyTS.

}
\seealso{
\code{\link{ggMyTS}}, \code{\link[reshape2]{melt.data.frame}}
}

