% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predProb.R
\name{predProb}
\alias{predProb}
\title{Predicted probabilities and confidence intervals from logit or probit model}
\usage{
predProb(model, predData, ci = TRUE, level = 0.95)
}
\arguments{
\item{model}{A \code{\link[stats]{glm}} model fit with \code{binomial} family
and either a \code{logit} or \code{probit} link function.}

\item{predData}{A data frame to pass to \code{\link[stats]{predict.glm}} in
which to look for variables with which to predict.}

\item{ci}{Logical value indicating whether to compute confidence intervals.}

\item{level}{The confidence level to use if \code{ci} is \code{TRUE}.}
}
\value{
A data frame with \code{predData} and the associated predicted
  probabilities. Confidence intervals are included if argument \code{ci} is
  \code{TRUE}.
}
\description{
Predicted probabilities and confidence intervals from logit or probit model
}
\examples{
GSS_2010$Y <-
  with(GSS_2010, cut(
    realinc,
    breaks = c(-Inf, median(realinc, na.rm = TRUE), Inf),
    labels = c("Low", "High")
  ))
logitmodel <- glm(Y ~ age + educ, data = GSS_2010, 
                  family = binomial(link = "logit"))
probitmodel <- glm(Y ~ age + educ, data = GSS_2010, 
                   family = binomial(link = "probit"))
predData <- data.frame(age = 20, educ = 15)

predProb(logitmodel, predData, ci = FALSE)
predProb(probitmodel, predData, ci= FALSE)
predData <- expand.grid(age = c(20, 50, 80), educ = c(5, 10, 15))
predProb(logitmodel, predData, ci = TRUE)
predProb(probitmodel, predData, ci= TRUE)

}

