% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propOddsTest.R
\name{propOddsTest}
\alias{propOddsTest}
\title{Test proportional odds assumption for ordinal logit models}
\usage{
propOddsTest(fit, relaxed.fit)
}
\arguments{
\item{fit}{An ordinal logit model fit with \code{\link[VGAM]{vglm}}
(\pkg{VGAM}) with \code{family = propodds}.}

\item{relaxed.fit}{An ordinal logit model fit with \code{\link[VGAM]{vglm}}
(\pkg{VGAM}) with \code{family = cumulative(reverse = TRUE)}.}
}
\value{
An object of class "htest".
}
\description{
Test proportional odds assumption for ordinal logit models
}
\examples{
\dontrun{
propOddsTest(fit, relaxed.fit)
}

}
\seealso{
\code{\link[VGAM]{vglm}}
}

