% Generated by roxygen2 (4.0.1): do not edit by hand
\name{propOddsTest}
\alias{propOddsTest}
\title{Test proportional odds assumption for ordinal logit models}
\usage{
propOddsTest(fit, relaxed.fit)
}
\arguments{
\item{fit}{An ordinal logit model fit with \code{vglm} (\pkg{VGAM}) with \code{family = propodds}}

\item{relaxed.fit}{An ordinal logit model fit with \code{vglm} (\pkg{VGAM}) with \code{family = cumulative(reverse = TRUE)}}
}
\value{
A list consisting of
\item{chi.sq}{The chi-squared test statistic}
\item{df}{The degrees of freedom parameter for the chi-squared distribution}
\item{p.value}{The associated p-value}
}
\description{
Test proportional odds assumption for ordinal logit models
}
\examples{
propOddsTest(fit, relaxed.fit)
}
\author{
Jonah Gabry <jsg2201@columbia.edu>
}
\seealso{
\code{\link{vglm}}
}

