% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ReverseThis}
\alias{ReverseThis}
\title{Reverse-code a variable}
\usage{
ReverseThis(var)
}
\arguments{
\item{var}{A numeric or factor variable.}
}
\value{
A reverse-coded version of \code{var}.
}
\description{
Reverse-code a variable
}
\examples{
x <- c(1,1,1,2,3,3)
ReverseThis(x)
}
\author{
Jonah Gabry <jsg2201@columbia.edu>
}

