% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rho.R
\name{rho}
\alias{rho}
\title{Compute fraction of variance due to u_i after fitting a model with
\code{lmer}}
\usage{
rho(fit)
}
\arguments{
\item{fit}{A model fit with \code{\link[lme4]{lmer}}.}
}
\value{
A number between 0 and 1. The fraction of variance due to u_i.
}
\description{
Compute fraction of variance due to u_i after fitting a model with
\code{lmer}
}
\details{
Note that in the output from \code{summary(fit)} the Std. Dev.
  column in the Random Effects section contains what STATA refers to as
  sigma_u and sigma_e. STATA also reports rho (fraction of variance due to
  u_i) but this is not given in the R output. The \code{rho} function
  computes this value.
}
\examples{
library(lme4)
fit <- lmer(Reaction ~ Days + (Days | Subject), sleepstudy)
rho(fit)

}

