% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmaRho.R
\name{sigmaRho}
\alias{sigmaRho}
\title{Compute sigma_u, sigma_e, and rho after fitting a fixed or random effects
model with \pkg{plm}}
\usage{
sigmaRho(fit)
}
\arguments{
\item{fit}{A model fit with \code{\link[plm]{plm}} and \code{model =
"within"} or \code{model = "random"}.}
}
\value{
A numeric vector containing the named components 
\itemize{ 
  \item{\code{sigma_u}}{Error due to differences between units.} 
  \item{\code{sigma_e}}{Error due to differences within units.} 
  \item{\code{rho}}{Proportion of variance due to unit effects.} 
} 
Results are also printed to the console.
}
\description{
Compute sigma_u, sigma_e, and rho after fitting a fixed or random effects
model with \pkg{plm}
}
\examples{
# Model from plm help page:
library(plm)
data("Produc")
fit <- plm(log(gsp) ~ log(pcap) + log(pc) + log(emp) + unemp,
           data = Produc, index = c("state","year"), model = "random")
sigmaRho(fit) 

}
\seealso{
\code{\link[plm]{plm}}
}

