% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stdCoef}
\alias{stdCoef}
\title{Compute standardized regression coefficients}
\usage{
stdCoef(fit)
}
\arguments{
\item{fit}{A model fit with \code[stats]{lm}.}
}
\value{
Standardized (beta) coefficients from a linear model.
}
\description{
Compute standardized regression coefficients
}
\examples{
fit <- lm(realinc ~ age + educ, data = GSS_2010)
stdCoef(fit)
}
\author{
Jonah Gabry <jsg2201@columbia.edu>
}

