% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TabX.R
\name{TabX}
\alias{TabX}
\title{Cross Tabulation}
\usage{
TabX(x, y, digits = 3)
}
\arguments{
\item{x, y}{The variables for the cross tabulation.}

\item{digits}{Number of digits for rounding proportions.}
}
\description{
\code{TabX} is a modified version of \code{\link[gmodels]{CrossTable}}
(\pkg{gmodels}). \code{TabX} will print a summary table with cell counts and
column proportions (similar to STATA's \code{tabulate} \emph{varname1}
\emph{varname2}, \code{col}).
}
\examples{
with(GSS_2010, TabX(sex, race))

}
\seealso{
\code{\link[gmodels]{CrossTable}}, \code{\link[stats]{xtabs}}, 
  \code{\link[base]{table}}, \code{\link[base]{prop.table}}
}

